// angular/src/app/models/UserCriteria.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v10.7.2.0 (Newtonsoft.Json v13.0.0.0) (http://NJsonSchema.org)
// </auto-generated>
//----------------------







export interface UserCriteria {
    userName: string | undefined;
    loginName: string | undefined;
    userMail: string | undefined;
    upi: string | undefined;
    roleId: number | undefined;
    ouId: number | undefined;
    isActive: boolean | undefined;
    createdBy: number | undefined;
    modifiedBy: number | undefined;
    createdDateFrom: Date | undefined;
    createdDateTo: Date | undefined;
    modifiedDateFrom: Date | undefined;
    modifiedDateTo: Date | undefined;
}