// angular/src/app/models/Role.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v10.7.2.0 (Newtonsoft.Json v13.0.0.0) (http://NJsonSchema.org)
// </auto-generated>
//----------------------







export enum RoleId {
    Initiator = 1,
    Advisor = 2,
    DecisionMaker = 3,
    Oec = 4,
    SupportGroup = 5,
    Viewer = 6,
    System = 7,
    AreaFunctionalAdmin = 8,
    AreaSecurityAdmin = 9,
    CountryFunctionalAdmin = 10,
    CountrySecurityAdmin = 11,
    GlobalFunctionalAdmin = 12,
    GlobalSecurityAdmin = 13,
    Notifier = 14,
    SecondAdvisor = 15,
    BusinessSupport = 16,
    Validator = 17,
    OecPlus = 18,
    Legal = 19,
    CountrySecurityLead = 20,
    PlaneaBusinessPlansUser = 21,
    PlaneaFinancialPlanning = 22,
    PlaneaConsultation = 23,
    PlaneaAdministrator = 24,
    ProductOwner = 25,
    CrossBorderAdmin = 26,
}

export enum PermissionId {
    Activities = 1,
    CreateMapFormsHome = 2,
    Searcher = 3,
    Dashboard = 4,
    BackofficeAdministration = 5,
    Delegations = 6,
    AdminUsers = 7,
    AdminOrganizationalUnits = 8,
    AdminMaintenance = 9,
    MyTasks = 10,
    AdminEvents = 11,
    AdminCustomers = 12,
    AdminDelegations = 13,
    AdminOds = 14,
    AdminDataExtraction = 15,
    AdminAuditTrail = 16,
    AdminProviders = 17,
    AdminCostCentersWbs = 18,
    AdminContracts = 19,
    AdminHcoSap = 20,
    AdminInternalOrders = 21,
    AdminLogs = 22,
    AdvancedSearch = 23,
    Events = 24,
    SecurityTracking = 25,
    Planea = 26,
    PlaneaGlobalAdminPlanea = 27,
    PlaneaSeeOnlyMyAccounts = 28,
    PlaneaSeeMedicalAmountsOnCommercial = 29,
    PlaneaTacticalPlansEdition = 30,
    PlaneaBudgetControlSeeAllBrands = 31,
    PlaneaAccountsConsultation = 32,
    PlaneaTacticalPlansConsultation = 33,
    PlaneaFixAnalysis = 34,
    PlaneaGlobalPlansAdmin = 35,
    PlaneaSipAccessWithoutMedicalDepartment = 36,
    PlaneaSipViewAllBrands = 37,
    PlaneaSipViewAllAccounts = 38,
    AdminTransparency = 39,
    CrossBorder = 40,
    AdminWaivers = 41,
    AdminMdm = 42,
}

export interface Role {
    roleId: RoleId;
    roleNameLocalizable: string | undefined;
    roleDescription: string | undefined;
    createdDate: Date;
    createdBy: number;
    modifiedDate: Date;
    modifiedBy: number;
    isActive: boolean;
    ordinal: number | undefined;
    permissionId: PermissionId;
    permissionNameLocalizable: string | undefined;
}